requirements.txt
fastapi
pydantic
pillow
opencv-python-headless
deepface
scikit-image
tf-keras
uvicorn